Sync latest EcoGPT observe runtime

+++ src/catalog/payments.ts
@@ -4,7 +4,7 @@ export const paymentEvents = {
 	payment_attempt: z.object({
 		event: z.literal('payment_attempt'),
 		payment: z.object({
-			amount_minor: z.number(),
+			amount_minor: z.number().int(),
 			currency: z.string(),
 			processor: z.string(),
 			status: z.string(),
